requirements.txt
boto3
PyYAML
psycopg2-binary
pymysql